refactor: migrate hotkey to svelte.ts and consolidate modal stores to modalState
- Rename hotkey.ts to hotkey.svelte.ts
- Replace openPresetList, openPersonaList, bookmarkListOpen with unified modalState $state object
