+++ src/ts/stores.svelte.ts
@@ -31,6 +31,12 @@ export const sideBarState = $state({
     closing: false
 })
 
+export const modalState = $state({
+    preset: false,
+    persona: false,
+    bookmark: false
+})
+
 export const selectedCharID = writable(-1)
 export const CurrentTriggerIdStore = writable<string | null>(null)
 export const CharEmotion = writable({} as {[key:string]: [string, string, number][]})
@@ -38,9 +44,6 @@ export const ViewBoxsize = writable({ width: 12 * 16, height: 12 * 16 }); // Def
 export const settingsOpen = writable(false)
 export const botMakerMode = writable(false)
 export const moduleBackgroundEmbedding = writable('')
-export const openPresetList = writable(false)
-export const openPersonaList = writable(false)
-export const bookmarkListOpen = writable(false)
 export const MobileGUI = writable(false)
 export const MobileGUIStack = writable(0)
 export const MobileSideBar = writable(0)
